refactor: move pointer related handlers to custom hook usePointerPlay
